backend/pyproject.toml
[tool.black]
extend-exclude = "alembic"
line-length = 79

[tool.poetry]
package-mode = false
name = "puzzle"
version = "0.1.0"
description = "Puzzle game"
authors = ["FinemechanicPub <93194456+FinemechanicPub@users.noreply.github.com>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
fastapi-users = "^13.0.0"
fastapi-users-db-sqlalchemy = "^6.0.1"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
asyncpg = "^0.29.0"
SQLAlchemy = "^2.0.31"
alembic = "^1.13.2"
pillow = "^10.4.0"
click = "^8.1.7"
pydantic-settings = "^2.4.0"
cachetools = "^5.4.0"

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.20.0"
flake8 = "^7.1.0"
black = "^24.8.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.0"
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
